requirements.txt
apache-airflow>=2.8.0
apache-airflow-providers-google>=10.0.0
pandas>=2.0.0
requests>=2.31.0
pytest>=7.0.0
